requirements.txt
django-admin-sortable
django-blog-zinnia
django-bootstrap-breadcrumbs>=0.7.0
django-bootstrap-form>=3.1
django-bootstrap3
django-compat
django-crispy-forms
django-debug-toolbar
django-npm
django-recaptcha
django-statictemplate
django-timezone-field>=2.0rc1
django<3
djmail>=0.10.0
geopy
markdown
pillow
psycopg2-binary>=2.5
requests
rules==1.4
transifex-client
typing; python_version < '3.5'

# From Prologin
git+https://github.com/prologin/django-massmailer
git+https://github.com/prologin/django-proloauth-client
